Replacing all along the project np.float to float.

+++ atom_evaluation/scripts/old/range_sensor_to_camera_evaluation_andre.py
@@ -79,8 +79,8 @@ def rangeToImage(collection, json_file, ss, ts, tf):
 
     # -- Project them to the image
     w, h = collection['data'][ts]['width'], collection['data'][ts]['height']
-    K = np.ndarray((3, 3), buffer=np.array(test_dataset['sensors'][ts]['camera_info']['K']), dtype=np.float)
-    D = np.ndarray((5, 1), buffer=np.array(test_dataset['sensors'][ts]['camera_info']['D']), dtype=np.float)
+    K = np.ndarray((3, 3), buffer=np.array(test_dataset['sensors'][ts]['camera_info']['K']), dtype=float)
+    D = np.ndarray((5, 1), buffer=np.array(test_dataset['sensors'][ts]['camera_info']['D']), dtype=float)
 
     pts_in_image, _, _ = projectToCamera(K, D, w, h, points_in_cam[0:3, :])
 
@@ -337,7 +337,7 @@ if __name__ == "__main__":
             cv2.waitKey()
 
     total_pts = len(delta_total)
-    delta_total = np.array(delta_total, np.float)
+    delta_total = np.array(delta_total, float)
     avg_error_x = np.sum(np.abs(delta_total[:, 0])) / total_pts
     avg_error_y = np.sum(np.abs(delta_total[:, 1])) / total_pts
     stdev = np.std(delta_total, axis=0)
